server/vcr-server/requirements.dev.txt
django_nose
black
flake8
